[multivariateexplorer] new class MultivariateExplorer
